feat(775): refactore sagas
