+++ src/core/sagas/fetchRequests/index.ts
@@ -1,0 +1,1 @@
+export {fetchCategoriesData} from './fetchCategoriesData';
