+++ src/core/sagas/filters/localFetchRequests/index.ts
@@ -1,0 +1,3 @@
+export {fetchCategorieData} from './fetchCategorieData';
+export {fetchFiltersData} from './fetchFiltersData';
+export {fetchRegionsData} from './fetchRegionsData';
